# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Bitácora SOC API
  description: API completa para el sistema de Bitácora SOC con gestión de entradas, checklist de turno, usuarios y reportes
  version: 1.0.0
  contact:
    name: Bitácora SOC Team
servers:
  - url: http://192.168.100.50:3000/api
    description: Servidor de desarrollo (IP)
  - url: http://localhost:3000/api
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, user, guest]
        isActive:
          type: boolean
        theme:
          type: string
          enum: [light, dark, sepia, pastel, cyberpunk]
        guestExpiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    Entry:
      type: object
      properties:
        _id:
          type: string
        content:
          type: string
        entryType:
          type: string
          enum: [operativa, incidente]
        entryDate:
          type: string
          format: date
        entryTime:
          type: string
          pattern: '^([01]\d|2[0-3]):([0-5]\d)$'
        tags:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdByUsername:
          type: string
        isGuestEntry:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      tags:
        - Usuarios
      summary: Listar usuarios
      description: Obtiene la lista de todos los usuarios (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    
    post:
      tags:
        - Usuarios
      summary: Crear usuario
      description: Crea un nuevo usuario (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - fullName
                - role
              properties:
                username:
                  type: string
                  example: jperez
                email:
                  type: string
                  example: jperez@example.com
                password:
                  type: string
                  example: password123
                fullName:
                  type: string
                  example: Juan Pérez
                role:
                  type: string
                  enum: [admin, user, guest]
                  example: user
      responses:
        '201':
          description: Usuario creado
        '400':
          description: Error de validación
...
  /entries:
    get:
      tags:
        - Entradas
      summary: Listar entradas
      description: Obtiene entradas con filtros y paginación
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
          description: Búsqueda de texto completo
        - in: query
          name: tags
          schema:
            type: string
          description: Tags separados por comas
          example: trellix,hunting
        - in: query
          name: entryType
          schema:
            type: string
            enum: [operativa, incidente]
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: userId
          schema:
            type: string
      responses:
        '200':
          description: Lista de entradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
    
    post:
      tags:
        - Entradas
      summary: Crear entrada
      description: Crea una nueva entrada en la bitácora
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - entryType
                - entryDate
                - entryTime
              properties:
                content:
                  type: string
                  example: Revisión de alertas en #Trellix. Todo operativo. #hunting
                entryType:
                  type: string
                  enum: [operativa, incidente]
                  example: operativa
                entryDate:
                  type: string
                  format: date
                  example: "2025-12-17"
                entryTime:
                  type: string
                  example: "14:30"
      responses:
        '201':
          description: Entrada creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  entry:
                    $ref: '#/components/schemas/Entry'
...
  /smtp:
    get:
      tags:
        - SMTP
      summary: Obtener configuración SMTP
      description: Obtiene la configuración del servidor de correo (solo admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuración SMTP
    
    post:
      tags:
        - SMTP
      summary: Configurar SMTP
      description: Crea o actualiza la configuración SMTP (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - username
                - password
                - host
                - port
                - useTLS
                - senderName
                - senderEmail
                - recipients
                - sendOnlyIfRed
              properties:
                provider:
                  type: string
                  enum: [office365, aws-ses, elastic-email, google-mail, google-workspace, mailgun, custom]
                username:
                  type: string
                password:
                  type: string
                host:
                  type: string
                port:
                  type: integer
                useTLS:
                  type: boolean
                senderName:
                  type: string
                senderEmail:
                  type: string
                recipients:
                  type: array
                  items:
                    type: string
                sendOnlyIfRed:
                  type: boolean
      responses:
        '200':
          description: Configuración guardada
...
  /config:
    get:
      tags:
        - Configuración
      summary: Obtener configuración
      description: Obtiene la configuración global de la aplicación
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuración actual
    
    put:
      tags:
        - Configuración
      summary: Actualizar configuración
      description: Actualiza la configuración global (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                guestModeEnabled:
                  type: boolean
                guestMaxDurationDays:
                  type: integer
                  minimum: 1
                  maximum: 30
                shiftCheckCooldownHours:
                  type: integer
                  minimum: 1
                  maximum: 1440
                logoUrl:
                  type: string
      responses:
        '200':
          description: Configuración actualizada
...